+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.1.1
+
+Added configuration parameters:
+  - `"clojureWarrior.bracketColors": ["#000", "#999", ...]`
+  - `"clojureWarrior.cycleBracketColors": true`
+  - `"clojureWarrior.misplacedBracketStyle": { "border": "2px solid #c33" }`
+
 # 0.1.0
 
 - Initial release